+++ lib/flack/app/executions.rb
@@ -4,6 +4,7 @@
 class Flack::App
 
   # GET /executions
+  #
   def get_executions(env)
 
 # TODO implement paging
@@ -11,6 +12,7 @@ class Flack::App
   end
 
   # GET /executions/<id>
+  #
   def get_executions_i(env)
 
     exe = @unit.executions[env['flack.args'][0]]
@@ -19,11 +21,48 @@ class Flack::App
   end
 
   # GET /executions/<exid>
+  # GET /executions/<domain>
+  # GET /executions/<domain>*
+  # GET /executions/<domain>.*
+  #
   def get_executions_s(env)
 
-    exe = @unit.executions.where(exid: env['flack.args'][0]).last
-    return respond_not_found(env) unless exe
-    respond(env, exe)
+    arg = env['flack.args'][0]
+
+    if arg.count('-') == 0
+      get_executions_by_domain(env, arg)
+    else
+      get_executions_by_exid(env, arg)
+    end
+  end
+
+  protected
+
+  def get_executions_by_exid(env, exid)
+
+    if exe = @unit.executions[exid: exid]
+      respond(env, exe)
+    else
+      respond_not_found(env)
+    end
+  end
+
+  def get_executions_by_domain(env, dom)
+
+    q =
+      @unit.executions
+    q =
+      if m = dom.match(/\A([^*]+)\*+\z/)
+        if m[1][-1, 1] == '.'
+          q.where(Sequel.like(:domain, "#{m[1]}%"))
+        else
+          q.where(Sequel[{ domain: m[1] }] | Sequel.like(:domain, "#{m[1]}.%"))
+        end
+      else
+        q.where(domain: dom)
+      end
+
+    respond(env, q.all)
   end
 end